requirements.txt
alabaster==0.7.12
altgraph==0.17
aniso8601==9.0.1
Babel==2.9.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
docutils==0.16
Flask-RESTful==0.3.8
future==0.18.2
idna==2.10
imagesize==1.2.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
pefile==2019.4.18
Pillow==8.2.0
Pygments==2.8.1
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
pyparsing==2.4.7
pyserial==3.5
python-dotenv==0.15.0
pytz==2021.1
pywin32-ctypes==0.2.0
requests==2.25.1
six==1.15.0
snowballstemmer==2.1.0
sockets==1.0.0
Sphinx==3.5.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
urllib3==1.26.4
waiting==1.4.1
Werkzeug==1.0.1
